ruff, test and offset implemented in q1asm
