+++ docs/releases/changelog-dev.md
@@ -2,6 +2,41 @@
 
 ### New features since last release
 
+- Extended `VariableExpression` capabilities (Qblox backend only)
+The capabilities of `VariableExpression` have been extended, and remain exclusive to the Qblox backend. 
+
+Previously, this type of expression was only supported in the Time Domain. It is now also available in the Voltage Domain, where it can be used to modify values in `qprogram` via the offset or the gain.
+The Time Domain behavior is unchanged. The updates described below therefore apply only to Voltage Domain operations.
+
+A combination of variables is now possible (Voltage Domain only), as shown below.
+  ```
+  qp = ql.Qprogram()
+  gain1 = qp.variable("gain1", ql.Domain.Voltage)
+  gain2 = qp.variable("gain2", ql.Domain.Voltage)
+  qp.set_gain("bus", gain1 + gain2)
+  
+  ```
+  These expressions are subject to some restrictions. A `NotImplementedError` is raised if any of these rules are violated:
+  - All constants must be integers.
+  - Expression chaining is not supported: at most two components (a variable and a constant, or two variables) are allowed. For example, the following code will raise a `NotImplementedError`:
+      ```
+    qp = ql.Qprogram(
+    gain1 = qp.variable("gain1", ql.Domain.Voltage)
+    qp.set_gain("bus", 10 + gain1 + 30)
+    )
+    ```
+  - Only addition and subtraction are supported. Any other type of operation will raise a `TypeError`.
+  - To facilitate the `Q1ASM` implementation, some expressions are internally reorganized in the `Variable` class without changing their semantics:
+    ```
+    gain + (-10) -> gain - abs(10)
+    - 10 + gain  -> gain - abs(10)
+    gain - (-10) -> gain + abs(10)
+    - gain       -> 0 - gain
+
+    ```
+
+  [#1057](https://github.com/qilimanjaro-tech/qililab/pull/1057)
+
 - This release introduces a significant architectural refactor of the digital and pulse-related layers, removes legacy dependencies, and aligns naming and abstractions with established superconducting-qubit literature.
 
   All references to **Qibo** have been removed from the codebase, along with the complete removal of the **pulse** module. The **digital** module has been fully restructured around a new, self-contained compilation and transpilation pipeline. This includes the introduction of a native **CircuitTranspiler** and **CircuitToQProgramCompiler**. The new **CircuitTranspiler** is responsible for decomposing circuits into the native gate set, managing logical and physical qubit layouts, and applying optimizations. It is implemented as a linear pipeline of `CircuitTranspilerPass` objects, replacing the previous Router, Placer, and Optimizer components that depended on Qibo-based implementations. Each transpiler pass now has a concrete, narrowly defined responsibility.
